+++ src/app/utils/item-utils.ts
@@ -18,7 +18,7 @@ import {
   killTrackerObjectivesByHash,
   killTrackerSocketTypeHash,
 } from 'app/search/d2-known-values';
-import { damageNamesByEnum } from 'app/search/search-filter-values';
+import { damageNamesByEnum, riteOfTheNineShinyWeapons } from 'app/search/search-filter-values';
 import modSocketMetadata, {
   ModSocketMetadata,
   modTypeTagByPlugCategoryHash,
@@ -160,6 +160,12 @@ export function itemCanBeEquippedByStoreId(
       // can be moved or is already here
       (!item.notransfer || item.owner === storeId) &&
       (allowPostmaster || !item.location.inPostmaster),
+  );
+}
+
+export function nonPullablePostmasterItem(item: DimItem): boolean {
+  return (
+    item.owner !== 'unknown' && !item.canPullFromPostmaster && Boolean(item.location.inPostmaster)
   );
 }
 
@@ -339,6 +345,11 @@ export function braveShiny(item: DimItem) {
     (s) =>
       s.plugOptions.some((s) => s.plugDef.plug.plugCategoryIdentifier === 'holofoil_skins_shared'), //
   );
+}
+
+/** Rite of the Nine "shiny" weapons with a unique appearance the the diagonal stripes */
+export function riteShiny(item: DimItem) {
+  return riteOfTheNineShinyWeapons.has(item.hash);
 }
 
 const ichToBreakerType = Object.entries(artifactBreakerMods).reduce<